+++ src/main/java/mx/unam/backend/config/WebMvcConfig.java
@@ -1,0 +1,22 @@
+package mx.unam.backend.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    /** {@inheritDoc} */
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        /**/
+
+        registry
+            .addMapping("/**")
+            .allowedOrigins("*")
+            .allowedMethods("GET", "POST","PUT", "DELETE", "OPTIONS", "HEAD");
+        /**/
+    }
+
+}
