--- src/main/java/unam/mx/backend/BackendApplication.java
@@ -1,18 +1,0 @@
-package unam.mx.backend;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@MapperScan("unam.mx.backend.mapper")
-@SpringBootApplication
-@EnableSwagger2
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-
-}
