--- src/main/java/unam/mx/backend/model/Rol.java
@@ -1,95 +1,0 @@
-/*
- *
- * Paquete:     unam.mx.backend.model
- * Proyecto:    licenciatura-backend
- * Tipo:        Class
- * Nombre:      Rol
- * Autor:       garellano
- * Versión:     1.0-SNAPSHOT
- *
- * Historia:
- *              Creación: 02 Mar 2022
- */
-package unam.mx.backend.model;
-/**
- * <p>Rol class.</p>
- *
- * @author garellano
- * @version $Id: $Id
- */
-public class Rol {
-    /*
-     * Atributos de la clase.
-     */
-    private int id;
-    private String nombre;
-    private boolean activo;
-
-    /**
-     * Constructor basado en los atributos de la clase.
-     *
-     * @param id a int.
-     * @param nombre a {@link java.lang.String} object.
-     * @param activo a boolean.
-     */
-    public Rol(int id, String nombre, boolean activo) {
-        this.id = id;
-        this.nombre = nombre;
-        this.activo = activo;
-    }
-
-    /**
-     * <p>Getter for the field <code>id</code>.</p>
-     *
-     * @return a int.
-     */
-    public int getId() {
-        return this.id;
-    }
-
-    /**
-     * <p>Setter for the field <code>id</code>.</p>
-     *
-     * @param id a int.
-     */
-    public void setId(int id) {
-        this.id = id;
-    }
-    
-    /**
-     * <p>Getter for the field <code>nombre</code>.</p>
-     *
-     * @return a {@link java.lang.String} object.
-     */
-    public String getNombre() {
-        return this.nombre;
-    }
-    
-    /**
-     * <p>Setter for the field <code>nombre</code>.</p>
-     *
-     * @param nombre a {@link java.lang.String} object.
-     */
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    /**
-     * <p>isActivo.</p>
-     *
-     * @return a boolean.
-     */
-    public boolean isActivo() {
-        return this.activo;
-    }
-
-    /**
-     * <p>Setter for the field <code>activo</code>.</p>
-     *
-     * @param activo a boolean.
-     */
-    public void setActivo(boolean activo) {
-        this.activo = activo;
-    }
-
-}
